Add KDE file handlers
Currently KDE will copy a media file into a temporary folder instead of trying to stream it if a KIO slave location file is started. This change will tell KDE to make mpv try to directly play the file. Perhaps the proper flags should be added according to the individual enabled features of the build but I suggest that be for the future.